Main merge tanya3

+++ src/main/java/sg/edu/iss/team6/repository/AdminRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import sg.edu.iss.team6.model.*;
 
 @Repository
-public interface AdminRepository extends JpaRepository<Admin, Integer> {
+public interface AdminRepository extends JpaRepository<Admin, Long> {
 
     @Query("SELECT a from Admin a where a.user.username = :username")
     public List<Admin>findByUser_Username(@Param("username")String username);
